+++ Projects/Apps/CordialDemo/CordialDemo/ViewControllers/Settings/SettingsViewController.swift
@@ -92,6 +92,14 @@ class SettingsViewController: UIViewController {
         alertController.addAction(customAction)
         alertController.addAction(cancelAction)
 
+        // iPadOS presentation
+        if let presenter = alertController.popoverPresentationController {
+            let button = sender as UIView
+            
+            presenter.sourceView = button
+            presenter.sourceRect = button.bounds
+        }
+        
         self.present(alertController, animated: true, completion: nil)
 
     }
@@ -181,8 +189,8 @@ class SettingsViewController: UIViewController {
     }
     
     func getQCSettings() -> Credentials {
-        let eventsStreamURL = "https://events-stream-svc.cordial-core.mobile-sdk-1-12-2.cordialdev.com/"
-        let messageHubURL = "https://message-hub-svc.cordial-core.mobile-sdk-1-12-2.cordialdev.com/"
+        let eventsStreamURL = "https://events-stream-svc.cordial-core.master.cordialdev.com/"
+        let messageHubURL = "https://message-hub-svc.cordial-core.master.cordialdev.com/"
         let accountKey = "qc-all-channels-cID-pk"
         let channelKey = "push"