+++ Sources/CordialSDK/CoreData/Helper/ContactOrderRequestsCoreData.swift
@@ -13,27 +13,68 @@ class ContactOrderRequestsCoreData {
     
     let entityName = "ContactOrderRequest"
     
-    func setContactOrderRequestsToCoreData(sendContactOrderRequests: [SendContactOrderRequest]) {
+    // MARK: Setting Data
+    
+    func putContactOrderRequests(sendContactOrderRequests: [SendContactOrderRequest]) {
         guard let context = CoreDataManager.shared.persistentContainer?.viewContext else { return }
         
         if let entity = NSEntityDescription.entity(forEntityName: self.entityName, in: context) {
-            sendContactOrderRequests.forEach { sendContactOrderRequest in
-                let newRow = NSManagedObject(entity: entity, insertInto: context)
+            for sendContactOrderRequest in sendContactOrderRequests {
+                let requestID = sendContactOrderRequest.order.orderID
                 
-                do {
-                    let sendContactOrderRequestData = try NSKeyedArchiver.archivedData(withRootObject: sendContactOrderRequest, requiringSecureCoding: true)
-                    
-                    newRow.setValue(sendContactOrderRequestData, forKey: "data")
-                    
-                    try context.save()
-                } catch let error {
-                    LoggerManager.shared.error(message: "CoreData Error: [\(error.localizedDescription)] Entity: [\(self.entityName)]", category: "CordialSDKCoreDataError")
+                guard let isContactOrderRequestExist = CoreDataManager.shared.isObjectExist(requestID: requestID, entityName: self.entityName) else { continue }
+                
+                if isContactOrderRequestExist {
+                    self.updateContactOrderRequest(sendContactOrderRequest: sendContactOrderRequest)
+                } else {
+                    let managedObject = NSManagedObject(entity: entity, insertInto: context)
+                    self.setContactOrderRequest(managedObject: managedObject, context: context, sendContactOrderRequest: sendContactOrderRequest)
                 }
             }
         }
     }
     
-    func getContactOrderRequestsFromCoreData() -> [SendContactOrderRequest] {
+    private func setContactOrderRequest(managedObject: NSManagedObject, context: NSManagedObjectContext, sendContactOrderRequest: SendContactOrderRequest) {
+        do {
+            let sendContactOrderRequestData = try NSKeyedArchiver.archivedData(withRootObject: sendContactOrderRequest, requiringSecureCoding: true)
+            
+            managedObject.setValue(sendContactOrderRequestData, forKey: "data")
+            managedObject.setValue(sendContactOrderRequest.order.orderID, forKey: "requestID")
+            managedObject.setValue(false, forKey: "flushing")
+            
+            CoreDataManager.shared.saveContext(context: context, entityName: self.entityName)
+            
+        } catch let error {
+            CoreDataManager.shared.deleteAll(entityName: self.entityName)
+            
+            LoggerManager.shared.error(message: "CoreData Error: [\(error.localizedDescription)] Entity: [\(self.entityName)]", category: "CordialSDKCoreDataError")
+        }
+    }
+    
+    private func updateContactOrderRequest(sendContactOrderRequest: SendContactOrderRequest) {
+        guard let context = CoreDataManager.shared.persistentContainer?.viewContext else { return }
+
+        let request = NSFetchRequest<NSFetchRequestResult>(entityName: self.entityName)
+        request.returnsObjectsAsFaults = false
+        
+        request.predicate = NSPredicate(format: "requestID = %@", sendContactOrderRequest.order.orderID)
+        
+        do {
+            guard let managedObjects = try context.fetch(request) as? [NSManagedObject] else { return }
+            
+            for managedObject in managedObjects {
+                self.setContactOrderRequest(managedObject: managedObject, context: context, sendContactOrderRequest: sendContactOrderRequest)
+            }
+        } catch let error {
+            CoreDataManager.shared.deleteAll(entityName: self.entityName)
+            
+            LoggerManager.shared.error(message: "CoreData Error: [\(error.localizedDescription)] Entity: [\(self.entityName)]", category: "CordialSDKCoreDataError")
+        }
+    }
+    
+    // MARK: Getting Data
+    
+    func fetchContactOrderRequests() -> [SendContactOrderRequest] {
         var sendContactOrderRequests = [SendContactOrderRequest]()
         
         guard let context = CoreDataManager.shared.persistentContainer?.viewContext else { return sendContactOrderRequests }
@@ -42,28 +83,54 @@ class ContactOrderRequestsCoreData {
         request.returnsObjectsAsFaults = false
         
         do {
-            let result = try context.fetch(request)
-            for managedObject in result as! [NSManagedObject] {
-                guard let anyData = managedObject.value(forKey: "data") else { continue }
-                let data = anyData as! Data
+            guard let managedObjects = try context.fetch(request) as? [NSManagedObject] else { return sendContactOrderRequests }
+            
+            for managedObject in managedObjects {
+                guard let data = managedObject.value(forKey: "data") as? Data else {
+                    CoreDataManager.shared.removeManagedObject(managedObject: managedObject, context: context, entityName: self.entityName)
+
+                    continue
+                }
                 
                 if let sendContactOrderRequest = try NSKeyedUnarchiver.unarchivedObject(ofClasses: [SendContactOrderRequest.self, Order.self, Address.self, CartItem.self] + API.DEFAULT_UNARCHIVER_CLASSES, from: data) as? SendContactOrderRequest,
                    !sendContactOrderRequest.isError {
                     
-                    sendContactOrderRequests.append(sendContactOrderRequest)
+                    guard let isFlushing = managedObject.value(forKey: "flushing") as? Bool else {
+                        CoreDataManager.shared.removeManagedObject(managedObject: managedObject, context: context, entityName: self.entityName)
+                        
+                        continue
+                    }
+                    
+                    if !isFlushing {
+                        managedObject.setValue(true, forKey: "flushing")
+                        
+                        sendContactOrderRequests.append(sendContactOrderRequest)
+                    }
                 } else {
-                    context.delete(managedObject)
-                    try context.save()
+                    CoreDataManager.shared.removeManagedObject(managedObject: managedObject, context: context, entityName: self.entityName)
                     
                     LoggerManager.shared.error(message: "Failed unarchiving SendContactOrderRequest", category: "CordialSDKError")
                 }
             }
+            
+            CoreDataManager.shared.saveContext(context: context, entityName: self.entityName)
+            
         } catch let error {
+            CoreDataManager.shared.deleteAll(entityName: self.entityName)
+            
             LoggerManager.shared.error(message: "CoreData Error: [\(error.localizedDescription)] Entity: [\(self.entityName)]", category: "CordialSDKCoreDataError")
         }
         
-        CoreDataManager.shared.deleteAllCoreDataByEntity(entityName: self.entityName)
-        
         return sendContactOrderRequests
+    }
+    
+    // MARK: Removing Data
+    
+    func removeContactOrderRequests(sendContactOrderRequests: [SendContactOrderRequest]) {
+        sendContactOrderRequests.forEach { sendContactOrderRequest in
+            let requestID = sendContactOrderRequest.order.orderID
+            
+            CoreDataManager.shared.removeObject(requestID: requestID, entityName: self.entityName)
+        }
     }
 }