+++ Sources/CordialSDK/CoreData/Helper/CustomEventRequestsCoreData.swift
@@ -13,91 +13,64 @@ class CustomEventRequestsCoreData {
     
     let entityName = "CustomEventRequest"
     
-    func putCustomEventRequestsToCoreData(sendCustomEventRequests: [SendCustomEventRequest]) {
-        guard let qtyCachedCustomEventRequests = self.getQtyCachedCustomEventRequests() else { return }
+    // MARK: Setting Data
+    
+    func putCustomEventRequests(sendCustomEventRequests: [SendCustomEventRequest]) {
+        guard let cachedCustomEventRequestsQty = self.fetchCachedCustomEventRequestsQty() else { return }
         
         let countSendCustomEventRequests = sendCustomEventRequests.count
         
-        let addingReporting = qtyCachedCustomEventRequests.addingReportingOverflow(countSendCustomEventRequests)
+        let addingReporting = cachedCustomEventRequestsQty.addingReportingOverflow(countSendCustomEventRequests)
         
         if !addingReporting.overflow {
             let sendCustomEventRequestsQty = addingReporting.partialValue
             
             if sendCustomEventRequestsQty > CordialApiConfiguration.shared.qtyCachedEventQueue {
                 let removeID = sendCustomEventRequestsQty - CordialApiConfiguration.shared.qtyCachedEventQueue
-                self.removeFirstCustomEventRequestsFromCoreData(removeTo: removeID)
+                self.removeFirstCustomEventRequests(removeTo: removeID)
             }
             
-            self.setCustomEventRequestsToCoreData(sendCustomEventRequests: sendCustomEventRequests)
+            self.setCustomEventRequests(sendCustomEventRequests: sendCustomEventRequests)
         }
     }
     
-    func removeCustomEventRequestsFromCoreData(sendCustomEventRequests: [SendCustomEventRequest]) {
-        sendCustomEventRequests.forEach { sendCustomEventRequest in
-            self.removeCustomEventRequestFromCoreData(sendCustomEventRequest: sendCustomEventRequest)
-        }
-    }
-    
-    func updateSendingCustomEventRequestsIfNeeded() {
-        if InternalCordialAPI().isUserLogin() {
-            DispatchQueue.main.async {
-                self.updateSendingCustomEventRequests()
-            }
-        }
-    }
-    
-    private func updateSendingCustomEventRequests() {
-        guard let context = CoreDataManager.shared.persistentContainer?.viewContext else { return }
-
-        let request = NSBatchUpdateRequest(entityName: self.entityName)
-        request.resultType = .statusOnlyResultType
-        
-        request.predicate = NSPredicate(format: "flushing = %@", NSNumber(value: true))
-        request.propertiesToUpdate = ["flushing": NSNumber(value: false)]
-        
-        do {
-            try context.execute(request)
-        } catch let error {
-            CoreDataManager.shared.deleteAllCoreDataByEntity(entityName: self.entityName)
-            
-            LoggerManager.shared.error(message: "CoreData Error: [\(error.localizedDescription)] Entity: [\(self.entityName)]", category: "CordialSDKCoreDataError")
-        }
-    }
-    
-    private func setCustomEventRequestsToCoreData(sendCustomEventRequests: [SendCustomEventRequest]) {
+    private func setCustomEventRequests(sendCustomEventRequests: [SendCustomEventRequest]) {
         guard let context = CoreDataManager.shared.persistentContainer?.viewContext else { return }
 
         if let entity = NSEntityDescription.entity(forEntityName: self.entityName, in: context) {
             for sendCustomEventRequest in sendCustomEventRequests {
-                guard let isCustomEventRequestExistAtCoreData = self.isCustomEventRequestExistAtCoreData(requestID: sendCustomEventRequest.requestID) else { continue }
+                let requestID = sendCustomEventRequest.requestID
                 
-                if isCustomEventRequestExistAtCoreData {
-                    self.updateCustomEventRequestsAtCoreData(sendCustomEventRequest: sendCustomEventRequest)
+                guard let isCustomEventRequestExist = CoreDataManager.shared.isObjectExist(requestID: requestID, entityName: self.entityName) else { continue }
+                
+                if isCustomEventRequestExist {
+                    self.updateCustomEventRequest(sendCustomEventRequest: sendCustomEventRequest)
                 } else {
                     let managedObject = NSManagedObject(entity: entity, insertInto: context)
-                    self.setCustomEventRequestToCoreData(managedObject: managedObject, context: context, sendCustomEventRequest: sendCustomEventRequest)
+                    self.setCustomEventRequest(managedObject: managedObject, context: context, sendCustomEventRequest: sendCustomEventRequest)
                 }
             }
         }
     }
     
-    private func setCustomEventRequestToCoreData(managedObject: NSManagedObject, context: NSManagedObjectContext, sendCustomEventRequest: SendCustomEventRequest) {
+    private func setCustomEventRequest(managedObject: NSManagedObject, context: NSManagedObjectContext, sendCustomEventRequest: SendCustomEventRequest) {
         do {
             let sendCustomEventRequestData = try NSKeyedArchiver.archivedData(withRootObject: sendCustomEventRequest, requiringSecureCoding: true)
             
             managedObject.setValue(sendCustomEventRequestData, forKey: "data")
             managedObject.setValue(sendCustomEventRequest.requestID, forKey: "requestID")
             managedObject.setValue(false, forKey: "flushing")
-
-            try context.save()
+            
+            CoreDataManager.shared.saveContext(context: context, entityName: self.entityName)
+            
         } catch let error {
-            CoreDataManager.shared.deleteAllCoreDataByEntity(entityName: self.entityName)
+            CoreDataManager.shared.deleteAll(entityName: self.entityName)
             
             LoggerManager.shared.error(message: "CoreData Error: [\(error.localizedDescription)] Entity: [\(self.entityName)]", category: "CordialSDKCoreDataError")
         }
     }
     
-    private func updateCustomEventRequestsAtCoreData(sendCustomEventRequest: SendCustomEventRequest) {
+    private func updateCustomEventRequest(sendCustomEventRequest: SendCustomEventRequest) {
         guard let context = CoreDataManager.shared.persistentContainer?.viewContext else { return }
 
         let request = NSFetchRequest<NSFetchRequestResult>(entityName: self.entityName)
@@ -109,63 +82,18 @@ class CustomEventRequestsCoreData {
             guard let managedObjects = try context.fetch(request) as? [NSManagedObject] else { return }
             
             for managedObject in managedObjects {
-                self.setCustomEventRequestToCoreData(managedObject: managedObject, context: context, sendCustomEventRequest: sendCustomEventRequest)
+                self.setCustomEventRequest(managedObject: managedObject, context: context, sendCustomEventRequest: sendCustomEventRequest)
             }
         } catch let error {
-            CoreDataManager.shared.deleteAllCoreDataByEntity(entityName: self.entityName)
+            CoreDataManager.shared.deleteAll(entityName: self.entityName)
             
             LoggerManager.shared.error(message: "CoreData Error: [\(error.localizedDescription)] Entity: [\(self.entityName)]", category: "CordialSDKCoreDataError")
         }
     }
     
-    private func removeCustomEventRequestFromCoreData(sendCustomEventRequest: SendCustomEventRequest) {
-        guard let context = CoreDataManager.shared.persistentContainer?.viewContext else { return }
-
-        let request = NSFetchRequest<NSFetchRequestResult>(entityName: self.entityName)
-        request.returnsObjectsAsFaults = false
-        
-        request.predicate = NSPredicate(format: "requestID = %@", sendCustomEventRequest.requestID)
-        
-        do {
-            guard let managedObjects = try context.fetch(request) as? [NSManagedObject] else { return }
-            
-            for managedObject in managedObjects {
-                CoreDataManager.shared.deleteManagedObjectByContext(managedObject: managedObject, context: context)
-            }
-        } catch let error {
-            CoreDataManager.shared.deleteAllCoreDataByEntity(entityName: self.entityName)
-            
-            LoggerManager.shared.error(message: "CoreData Error: [\(error.localizedDescription)] Entity: [\(self.entityName)]", category: "CordialSDKCoreDataError")
-        }
-    }
+    // MARK: Getting Data
     
-    private func removeFirstCustomEventRequestsFromCoreData(removeTo id: Int) {
-        guard let context = CoreDataManager.shared.persistentContainer?.viewContext else { return }
-
-        let request = NSFetchRequest<NSFetchRequestResult>(entityName: self.entityName)
-        request.returnsObjectsAsFaults = false
-
-        do {
-            guard let managedObjects = try context.fetch(request) as? [NSManagedObject] else { return }
-            
-            var countId = 0
-            for managedObject in managedObjects {
-                if id > countId {
-                    countId+=1
-                    
-                    CoreDataManager.shared.deleteManagedObjectByContext(managedObject: managedObject, context: context)
-                } else {
-                    break
-                }
-            }
-        } catch let error {
-            CoreDataManager.shared.deleteAllCoreDataByEntity(entityName: self.entityName)
-            
-            LoggerManager.shared.error(message: "CoreData Error: [\(error.localizedDescription)] Entity: [\(self.entityName)]", category: "CordialSDKCoreDataError")
-        }
-    }
-    
-    func getCustomEventRequestsFromCoreData() -> [SendCustomEventRequest] {
+    func fetchCustomEventRequests() -> [SendCustomEventRequest] {
         var sendCustomEventRequests = [SendCustomEventRequest]()
         
         guard let context = CoreDataManager.shared.persistentContainer?.viewContext else { return sendCustomEventRequests }
@@ -178,7 +106,7 @@ class CustomEventRequestsCoreData {
             
             for managedObject in managedObjects {
                 guard let data = managedObject.value(forKey: "data") as? Data else {
-                    CoreDataManager.shared.deleteManagedObjectByContext(managedObject: managedObject, context: context)
+                    CoreDataManager.shared.removeManagedObject(managedObject: managedObject, context: context, entityName: self.entityName)
 
                     continue
                 }
@@ -187,58 +115,35 @@ class CustomEventRequestsCoreData {
                    !sendCustomEventRequest.isError {
                     
                     guard let isFlushing = managedObject.value(forKey: "flushing") as? Bool else {
-                        CoreDataManager.shared.deleteManagedObjectByContext(managedObject: managedObject, context: context)
+                        CoreDataManager.shared.removeManagedObject(managedObject: managedObject, context: context, entityName: self.entityName)
                         
                         continue
                     }
                     
                     if !isFlushing {
                         managedObject.setValue(true, forKey: "flushing")
-                        try context.save()
                         
                         sendCustomEventRequests.append(sendCustomEventRequest)
                     }
                 } else {
-                    CoreDataManager.shared.deleteManagedObjectByContext(managedObject: managedObject, context: context)
+                    CoreDataManager.shared.removeManagedObject(managedObject: managedObject, context: context, entityName: self.entityName)
                     
                     LoggerManager.shared.error(message: "Failed unarchiving SendCustomEventRequest", category: "CordialSDKError")
                 }
             }
+            
+            CoreDataManager.shared.saveContext(context: context, entityName: self.entityName)
+            
         } catch let error {
-            CoreDataManager.shared.deleteAllCoreDataByEntity(entityName: self.entityName)
+            CoreDataManager.shared.deleteAll(entityName: self.entityName)
             
             LoggerManager.shared.error(message: "CoreData Error: [\(error.localizedDescription)] Entity: [\(self.entityName)]", category: "CordialSDKCoreDataError")
         }
 
         return sendCustomEventRequests
     }
-    
-    private func isCustomEventRequestExistAtCoreData(requestID: String) -> Bool? {
-        guard let context = CoreDataManager.shared.persistentContainer?.viewContext else { return nil }
 
-        let request = NSFetchRequest<NSFetchRequestResult>(entityName: self.entityName)
-        request.returnsObjectsAsFaults = false
-
-        let predicate = NSPredicate(format: "requestID = %@", requestID)
-        request.predicate = predicate
-        
-        do {
-            if let managedObjects = try context.fetch(request) as? [NSManagedObject],
-               managedObjects.count > 0 {
-                
-                return true
-            }
-        } catch let error {
-            CoreDataManager.shared.deleteAllCoreDataByEntity(entityName: self.entityName)
-            
-            LoggerManager.shared.error(message: "CoreData Error: [\(error.localizedDescription)] Entity: [\(self.entityName)]", category: "CordialSDKCoreDataError")
-        }
-
-        return false
-        
-    }
-
-    func getQtyCachedCustomEventRequests() -> Int? {
+    func fetchCachedCustomEventRequestsQty() -> Int? {
         guard let context = CoreDataManager.shared.persistentContainer?.viewContext else { return nil }
 
         let request = NSFetchRequest<NSFetchRequestResult>(entityName: self.entityName)
@@ -250,11 +155,48 @@ class CustomEventRequestsCoreData {
             return count
             
         } catch let error {
-            CoreDataManager.shared.deleteAllCoreDataByEntity(entityName: self.entityName)
+            CoreDataManager.shared.deleteAll(entityName: self.entityName)
             
             LoggerManager.shared.error(message: "CoreData Error: [\(error.localizedDescription)] Entity: [\(self.entityName)]", category: "CordialSDKCoreDataError")
         }
         
         return nil
     }
+    
+    // MARK: Removing Data
+    
+    func removeCustomEventRequests(sendCustomEventRequests: [SendCustomEventRequest]) {
+        sendCustomEventRequests.forEach { sendCustomEventRequest in
+            let requestID = sendCustomEventRequest.requestID
+            
+            CoreDataManager.shared.removeObject(requestID: requestID, entityName: self.entityName)
+        }
+    }
+    
+    private func removeFirstCustomEventRequests(removeTo id: Int) {
+        guard let context = CoreDataManager.shared.persistentContainer?.viewContext else { return }
+
+        let request = NSFetchRequest<NSFetchRequestResult>(entityName: self.entityName)
+        request.returnsObjectsAsFaults = false
+
+        do {
+            guard let managedObjects = try context.fetch(request) as? [NSManagedObject] else { return }
+            
+            var countId = 0
+            for managedObject in managedObjects {
+                if id > countId {
+                    countId += 1
+                    
+                    CoreDataManager.shared.removeManagedObject(managedObject: managedObject, context: context, entityName: self.entityName)
+                } else {
+                    break
+                }
+            }
+        } catch let error {
+            CoreDataManager.shared.deleteAll(entityName: self.entityName)
+            
+            LoggerManager.shared.error(message: "CoreData Error: [\(error.localizedDescription)] Entity: [\(self.entityName)]", category: "CordialSDKCoreDataError")
+        }
+    }
+
 }