+++ Sources/CordialSDK/CoreData/ManagedObjects/ContactLogout+CoreDataClass.swift
@@ -2,8 +2,8 @@
 //  ContactLogout+CoreDataClass.swift
 //  CordialSDK
 //
-//  Created by Yan Malinovsky on 5/20/19.
-//  Copyright © 2019 cordial.io. All rights reserved.
+//  Created by Yan Malinovsky on 27.07.2023.
+//  Copyright © 2023 cordial.io. All rights reserved.
 //
 //