+++ Sources/CordialSDK/Endpoints/UpsertContactCart/Helper/ContactCartSender.swift
@@ -18,12 +18,12 @@ class ContactCartSender {
             if ReachabilityManager.shared.isConnectedToInternet {
                 self.upsertContactCartData(upsertContactCartRequest: upsertContactCartRequest)
             } else {
-                CoreDataManager.shared.contactCartRequest.setContactCartRequestToCoreData(upsertContactCartRequest: upsertContactCartRequest)
+                CoreDataManager.shared.contactCartRequest.putContactCartRequest(upsertContactCartRequest: upsertContactCartRequest)
                 
                 LoggerManager.shared.info(message: "Sending contact cart failed. Saved to retry later. Request ID: [\(upsertContactCartRequest.requestID)] Error: [No Internet connection]", category: "CordialSDKUpsertContactCart")
             }
         } else {
-            CoreDataManager.shared.contactCartRequest.setContactCartRequestToCoreData(upsertContactCartRequest: upsertContactCartRequest)
+            CoreDataManager.shared.contactCartRequest.putContactCartRequest(upsertContactCartRequest: upsertContactCartRequest)
             
             LoggerManager.shared.info(message: "Sending contact cart failed. Saved to retry later. Request ID: [\(upsertContactCartRequest.requestID)] Error: [User no login]", category: "CordialSDKUpsertContactCart")
         }
@@ -48,16 +48,20 @@ class ContactCartSender {
     }
     
     func completionHandler(upsertContactCartRequest: UpsertContactCartRequest) {
+        CoreDataManager.shared.contactCartRequest.removeContactCartRequest(upsertContactCartRequest: upsertContactCartRequest)
+        
         LoggerManager.shared.info(message: "Contact cart has been sent. Request ID: [\(String(upsertContactCartRequest.requestID))]", category: "CordialSDKUpsertContactCart")
     }
     
     func systemErrorHandler(upsertContactCartRequest: UpsertContactCartRequest, error: ResponseError) {
-        CoreDataManager.shared.contactCartRequest.setContactCartRequestToCoreData(upsertContactCartRequest: upsertContactCartRequest)
+        CoreDataManager.shared.contactCartRequest.putContactCartRequest(upsertContactCartRequest: upsertContactCartRequest)
         
         LoggerManager.shared.info(message: "Sending contact cart failed. Saved to retry later. Request ID: [\(upsertContactCartRequest.requestID)] Error: [\(error.message)]", category: "CordialSDKUpsertContactCart")
     }
     
     func logicErrorHandler(upsertContactCartRequest: UpsertContactCartRequest, error: ResponseError) {
+        CoreDataManager.shared.contactCartRequest.removeContactCartRequest(upsertContactCartRequest: upsertContactCartRequest)
+        
         NotificationCenter.default.post(name: .cordialUpsertContactCartLogicError, object: error)
         
         LoggerManager.shared.error(message: "Sending contact cart failed. Will not retry. Request ID: [\(upsertContactCartRequest.requestID)] Error: [\(error.message)]", category: "CordialSDKUpsertContactCart")