+++ Sources/CordialSDK/Swizzler/CordialSwizzlerHelper.swift
@@ -28,8 +28,8 @@ class CordialSwizzlerHelper {
         
         if let mcID = pushNotificationParser.getMcID(userInfo: userInfo) {
             if pushNotificationParser.isPayloadContainInboxMessage(userInfo: userInfo) {
-                CoreDataManager.shared.inboxMessagesCache.removeInboxMessageFromCoreData(mcID: mcID)
-                CoreDataManager.shared.inboxMessagesContent.removeInboxMessageContentFromCoreData(mcID: mcID)
+                CoreDataManager.shared.inboxMessagesCache.removeInboxMessage(mcID: mcID)
+                CoreDataManager.shared.inboxMessagesContent.removeInboxMessageContent(mcID: mcID)
                 
                 // UIKit
                 if let inboxMessageDelegate = CordialApiConfiguration.shared.inboxMessageDelegate {